+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 3.2.04
+* Adds chainable ifSuccess and ifFailure consumer methods to the response. Now you can handle the response like
+```java
+ Unirest.get("https://localhost/somewhere)
+                .asString()
+                .ifSuccess(r -> log.info("Yippy!"))
+                .ifFailure(r -> log.error("Booo"));
+```
+
 # 3.2.03
 * Allow the configuration of header suppliers.
 ```java